Ignore java gradle worktree changes (#24)
* Added Java SDK to build process

The build and test workflow has been updated to include the Java SDK. This change ensures that the Java SDK is built alongside the Go, Node.js, and Python SDKs during the CI/CD process.

* Updated GitHub workflows

Added a step to display the git status in the 'build-test' workflow. This will help in debugging by providing more visibility into the state of the repository at different stages of the build process.